+++ Commands/basicCommands.py
@@ -6,6 +6,7 @@ log = logging.getLogger(__name__)
 
 from subprocess import call									## System module
 from os import _exit, getpid								## System module
+from platform import system									## System module
 from random import randint									## System module
 
 import Functions.basicData as bd							## Own module
@@ -24,9 +25,16 @@ def restartP(bot, update):
 	bd.startWithCommand(bot, update)
 
 	if bd.user_id == bd.chatIDDeveloper:
-		bot.sendMessage(chat_id=bd.chat_id, text=ms.restarting, reply_to_message_id=bd.message.message_id)
-		call("./startBot.sh " + str(getpid()), shell=True)
+		if system() == "Linux":
+			try:
+				bot.sendMessage(chat_id=bd.chat_id, text=ms.restarting, reply_to_message_id=bd.message.message_id)
+				call("./startBot.sh " + str(getpid()), shell=True)
+			except Exception as e:
+				log.error(str(e))
+				bot.sendMessage(chat_id=bd.chat_id, text=ms.errorExecCommandUser, reply_to_message_id=bd.message.message_id)
 
+		else:
+			bot.sendMessage(chat_id=bd.chat_id, text=ms.restartWrongOS, reply_to_message_id=bd.message.message_id)
 	else:
 		bot.sendMessage(chat_id=bd.chat_id, text=ms.notAdmin[randint(0, len(ms.notAdmin)-1)], reply_to_message_id=bd.message.message_id)
 
@@ -38,24 +46,6 @@ def stopP(bot, update):
 	if bd.user_id == bd.chatIDDeveloper:
 		bot.sendMessage(chat_id=bd.chat_id, text=ms.stopping, reply_to_message_id=bd.message.message_id)
 		_exit(1)
-
-	else:
-		bot.sendMessage(chat_id=bd.chat_id, text=ms.notAdmin[randint(0, len(ms.notAdmin)-1)], reply_to_message_id=bd.message.message_id)
-
-
-# Command /updateP (Private)
-def updateP(bot, update):
-	bd.startWithCommand(bot, update)
-
-	if bd.user_id == bd.chatIDDeveloper:
-		bot.sendMessage(chat_id=bd.chatIDDeveloper, text=ms.updating, reply_to_message_id=bd.message.message_id)
-		call("wget -qP /$HOME/BirthdayBot/ https://api.github.com/repos/javigines/EventsBot-CoreDumped/tarball/master", shell=True)
-		call("tar -xzf /$HOME/BirthdayBot/master -C $HOME", shell=True)
-		call("rm -f /$HOME/BirthdayBot/master*", shell=True)
-		call("cp -rf $HOME/javigines-EventsBot-CoreDumped-*/* $HOME/BirthdayBot/ ", shell=True)
-		call("rm -rf $HOME/javigines-EventsBot-CoreDumped-*/", shell=True)
-
-		bot.sendMessage(chat_id=bd.chatIDDeveloper, text=ms.updateDone)
 
 	else:
 		bot.sendMessage(chat_id=bd.chat_id, text=ms.notAdmin[randint(0, len(ms.notAdmin)-1)], reply_to_message_id=bd.message.message_id)
@@ -82,7 +72,7 @@ def changelog(bot, update):
 		changelog = open("CHANGELOG.md", mode="r")
 		changelogTemp = changelog.read()
 		changelog.close()
-		changelogTemp = "## [" + changelogTemp.split("## [")[2] + "## ["+ changelogTemp.split("## [")[1]+"\n\nPara ver el changelog completo:\nhttps://goo.gl/eHrop3"
+		changelogTemp = ms.changelogMessage.replace('$args1', "## [" + changelogTemp.split("## [")[2]).replace('$args2', "## ["+ changelogTemp.split("## [")[1])
 		changelogTemp = changelogTemp.split("\n")
 		for line in changelogTemp:
 			if "###" in line:
@@ -96,20 +86,6 @@ def changelog(bot, update):
 		bot.sendMessage(chat_id=bd.chat_id, text=ms.groupChangelogUser, reply_to_message_id=bd.message.message_id)
 
 
-# Changelog command /speak
-def speak(bot, update, args):
-	bd.startWithCommand(bot, update)
-
-	if bd.user_id == bd.chatIDDeveloper:
-		try:
-			bot.sendMessage(chat_id=args[0], text=' '.join(args).split('|')[1])
-			bot.sendMessage(chat_id=bd.chat_id, text=ms.messageSend, reply_to_message_id=bd.message.message_id)
-		except:
-			bot.sendMessage(chat_id=bd.chat_id, text=ms.incorrectChatId, reply_to_message_id=bd.message.message_id)
-
-	else:
-		bot.sendMessage(chat_id=bd.chat_id, text=ms.notAdmin[randint(0, len(ms.notAdmin)-1)], reply_to_message_id=bd.message.message_id)
-
 # Changelog command /contact
 def contact(bot, update, args):
 	bd.startWithCommand(bot, update)
@@ -120,24 +96,6 @@ def contact(bot, update, args):
 	except Exception as e:
 		log.error(str(e))
 		bot.sendMessage(chat_id=bd.chat_id, text=ms.errorExecCommandUser, reply_to_message_id=bd.message.message_id)
-
-# Download file command /downloadp (Private)
-def downloadP(bot, update, args):
-	bd.startWithCommand(bot, update, args)
-
-	if bd.user_id == bd.chatIDDeveloper:
-		try:
-			bot.sendMessage(chat_id=bd.chat_id, text=ms.downloadInProgress)
-			fileDocument = open("".join(args), mode="rb")
-			bot.sendDocument(chat_id=bd.chatIDDeveloper, document=fileDocument, reply_to_message_id=bd.message.message_id)
-			fileDocument.close()
-			bot.sendMessage(chat_id=bd.chat_id, text=ms.downloadComplete)
-		except Exception as e:
-			log.error(str(e))
-			bot.sendMessage(chat_id=bd.chat_id, text=ms.errorExecCommandUser, reply_to_message_id=bd.message.message_id)
-
-	else:
-		bot.sendMessage(chat_id=bd.chat_id, text=ms.notAdmin[randint(0, len(ms.notAdmin)-1)], reply_to_message_id=bd.message.message_id)
 
 
 log.info('BasicCommands Module Loaded.')